requirements.txt
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
docker-pycreds==0.4.0
gitdb==4.0.10
GitPython==3.1.31
hydra-core==1.3.2
idna==3.4
omegaconf==2.3.0
packaging==23.0
pathtools==0.1.2
-e git+ssh://git@github.com/BrunoKM/phenomenological-bayes.git@3dd31da69b9e104e1d84438be25dda3584b60fca#egg=PhenomBayes
protobuf==4.22.0
psutil==5.9.4
PyYAML==6.0
requests==2.28.2
sentry-sdk==1.16.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
urllib3==1.26.14
wandb==0.13.10
